@@ -1,9 +1,10 @@
 {
     "name": "@semivan/eslint-config-ts",
-    "version": "1.2.6",
+    "version": "2.0.1",
     "description": "TypeScript ESLint config",
     "author": "Ivan Semenikhin <semenihin.ivan@gmail.com>",
     "license": "MIT",
+    "type": "module",
     "exports": {
         ".": "./src/index.js",
         "./type-checked": "./src/type-checked.js",
@@ -17,14 +18,14 @@
         "url": "https://github.com/semivan/javascript"
     },
     "dependencies": {
-        "@semivan/eslint-config": "^1.2.8"
+        "@semivan/eslint-config": "^2.0.1"
     },
     "devDependencies": {
-        "@stylistic/eslint-plugin": "^1.7.0",
-        "@typescript-eslint/eslint-plugin": "^7.3.1",
-        "@typescript-eslint/parser": "^7.3.1",
-        "eslint": "^8.57.0",
-        "typescript": "^5.4.2"
+        "@stylistic/eslint-plugin": "^2.12.1",
+        "@typescript-eslint/eslint-plugin": "^8.18.0",
+        "@typescript-eslint/parser": "^8.18.0",
+        "eslint": "^9.16.0",
+        "typescript": "^5.7.2"
     },
     "peerDependencies": {
         "@stylistic/eslint-plugin": "*",
